Prepare package installation for different distros
So far, the package installation role takes care of the packages that I
have installed on my Arch machine. But I want to be able to use it for
different distros. Therefore I moved everything arch related into
separate files that have "Archlinux" in there names, so they can be
included based on the ansible_os_family variable.