[WIP] Install neovim version 0.7.0 or higher
So, I think it is for treesitter, version 0.7.0 of neovim is needed. As
of now, the version of neovim in the apt packages is lower than that.
Therefore, I played around to check the currently installed version of
neovim and if it is too low, I wanted to download a released deb package
with a version that is high enough, and blablabla.

While working on it, I found out that the version of neovim in the snap
package repository is high enough, and just installed that.

But I didn't want to loose what I had done, because it might be useful
in the future. That's why I just checked it in on a separate branch for
future reference.