Install pip packages depending on super group
So for example, if the **multimedia apt** packages are installed,
it makes sense to automatically install the **multimedia pip** as well.
That's what this commit does. The same is actually also true for
additional snap packages, but with snap, I'm not sure if it would make
sense to forbit them on some systems entirely. In such a case you would
just have to live with the fact that e.g  alacritty is not available.