Install xmonad on Arch Linux
Xmaond is finally available in the Arch package repo in a version that I
need, therefore I can just install it as a pacman package. Unfortuantely
on Ubunut the version in the apt repo is not high enough. On Ubuntu, I
will have to build xmoand from source. But I will probably not do that
until I really need it. Maybe by then, the official repos will have a
version that is sufficient for me (I need at least 0.17).