Remove xmonad recompile task
This task caused problems when testing it with molecule, because my user
configuration with the .config/xmonad/xmonad.hs does not exist in this
case. After realizing that, I also thought that this does not belong
into the package installtion role. It should be done as a separate step
in the playbook that used this role.